Make the sizes attribute opaque
So that we do one less copying.  To enable this, the type
of `PointCloud::sizes` is now `std::vector<double>` so that
it matches the type of `float` in Python (which is 64 bits).

As a result,
```
pcd.sizes = nd.array(..)
```
no longer works.  We need to cast a `numpy` array to type
`o3d.utility.DoubleVector`.

```
pcd.sizes = o3d.utility.DoubleVector(np.array(..))
```

+++ cpp/open3d/geometry/PointCloud.h
@@ -384,7 +384,7 @@ public:
     /// RGB colors of points.
     std::vector<Eigen::Vector3d> colors_;
     /// Sizes of points.
-    std::vector<float> sizes_;
+    std::vector<double> sizes_;
 };
 
 }  // namespace geometry
